+++ pagegraph/graph/requests.py
@@ -1,5 +1,6 @@
+from enum import StrEnum
 from dataclasses import dataclass, field
-from typing import cast, TYPE_CHECKING, Union
+from typing import cast, Optional, TYPE_CHECKING, Union
 
 from pagegraph.types import RequestId, Url, PageGraphEdgeId
 from pagegraph.serialize import Reportable, RequestCompleteReport
@@ -11,6 +12,43 @@ if TYPE_CHECKING:
     from pagegraph.graph.edge import RequestStartEdge, RequestRedirectEdge
     from pagegraph.graph.edge import RequestResponseEdge, RequestCompleteEdge
     from pagegraph.graph.edge import RequestErrorEdge
+
+
+# Values are defined by Blink, in `Resource::ResourceTypeToString`.
+# See third_party/blink/renderer/platform/loader/fetch/resource.h.
+# The OTHER catch all case covers the additional types
+# defined in `blink::Resource::InitiatorTypeNameToString`.
+class ResourceType(StrEnum):
+    ATTRIBUTION_RESOURCE = "Attribution resource"
+    AUDIO = "Audio"
+    CSS_RESOURCE = "CSS resource"
+    CSS_RESOURCE_UA = "User Agent CSS resource"
+    CSS_STYLESHEET = "CSS stylesheet"
+    DICTIONARY = "Dictionary"
+    DOCUMENT = "Document"
+    FETCH = "Fetch"
+    FONT = "Font"
+    ICON = "Icon"
+    IMAGE = "Image"
+    INTERNAL_RESOURCE = "Internal resource"
+    LINK_ELM_RESOURCE = "Link element resource"
+    LINK_PREFETCH = "Link prefetch resource"
+    MANIFEST = "Manifest"
+    MOCK = "Mock"
+    PROCESSING_INSTRUCTION = "Processing instruction"
+    RAW = "Raw"
+    REQUEST = "Request"
+    SCRIPT = "Script"
+    SPECULATION_RULE = "SpeculationRule"
+    SVG = "SVG document"
+    SVG_USE_ELM_RESOURCE = "SVG Use element resource"
+    TEXT_TRACK = "Text track"
+    TRACK = "Track"
+    VIDEO = "Video"
+    XML_HTTP_REQUEST = "XMLHttpRequest"
+    XML_RESOURCE = "XML resource"
+    XSL_STYLESHEET = "XSL stylesheet"
+    OTHER = "Other"  # Fallback / catchall case
 
 
 @dataclass
@@ -59,6 +97,24 @@ class RequestChain(Reportable):
         if self.result and self.result.is_request_complete_edge():
             return cast("RequestCompleteEdge", self.result).hash()
         return None
+
+    def success_request(self) -> Optional["RequestCompleteEdge"]:
+        if self.result and self.result.is_request_complete_edge():
+            return cast("RequestCompleteEdge", self.result)
+        return None
+
+    def error_request(self) -> Optional["RequestErrorEdge"]:
+        if self.result and self.result.is_request_error_edge():
+            return cast("RequestErrorEdge", self.result)
+        return None
+
+    def resource_type(self) -> ResourceType:
+        return self.request.resource_type()
+
+    def final_url(self) -> Url:
+        if len(self.redirects) == 0:
+            return self.request.url()
+        return self.redirects[-1].url()
 
 
 def request_chain_for_edge(request_edge: "RequestStartEdge") -> RequestChain: